requirements.txt
flask
openai
flask-cors
flask-socketio
python-dotenv